refactor: modify naming
