+++ tomcat/src/main/java/com/techcourse/restController/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.restController;
 
 import com.techcourse.service.UserService;
